sysfs: use poll implementation for WithStdin
If we pass os.Stdin to WithStdin, a special code path is used which implements
Poll() with os-specific Poll APIs. This is used in poll_oneoff to implement
asynchronous I/O with stdin as a special case:

see wazero/internal/sys/fs.go

- InitFSContext(stdin, stdout, ...)
  - inFile, err := stdinFileEntry(stdin)
    - if nil, no-op file descriptor
    - if *os.File, uses NewStdioFile(true, f)
    - otherwise returns &StdinFile{Reader: r}

If we want to implement Poll with a custom io.Reader passed to WithStdin, this
is currently not possible.

This commit uses Poll on the io.Reader passed to WithStdin if present.

See: https://github.com/tetratelabs/wazero/issues/1500#issuecomment-3041125375

Signed-off-by: Christian Stewart <christian@aperture.us>